Addressed arithmetic overflow and other code analysis warnings
* Initialized PipelineMessage index
* Marked a possible coding error in ImportAsciiDataArray where two identical error conditions are checked using || operator.
* Fixed an incorrect cast in ReadASCIIData where a double is cast to a float for a method parameter that is expecting a double.  Due to multiple calculations of the same value, it was saved to a constant before the first calculation was required.
* Changed a for loop variable in geometry compute methods to avoid default castings when dereferencing a pointer multiple times per iteration.
* Changed geometry requirements in addAttributeMatrix from using || for enum checks to &&.  The previous versions could never be false and invalidated the entire method.
* Fixed a possible overflow when calculating ImageGeom and RectGridGeom coords.  The previous versions would have overflowed before casting to a larger size.
* Updated variables used to store min and max float values to constexpr to take advantage of the fact that they are known at compile time and are never modified.
* Removed unnecessary casts in SIMPLibRandom where doubles were cast to float and then saved as a double.
* Fixed ColorTable::GetColorTable due to the current implementation going out of array bounds.  The 2D array float color[8][3] was being accessed with indices [8][2].

+++ Source/SIMPLib/Geometry/IGeometry.cpp
@@ -422,7 +422,7 @@ void IGeometry::sendThreadSafeProgressMessage(int64_t counter, int64_t max)
 
   if(m_ProgressCounter > prog)
   {
-    int64_t progressInt = static_cast<int64_t>((static_cast<float>(m_ProgressCounter) / max) * 100.0f);
+    int64_t progressInt = static_cast<int64_t>((static_cast<double>(m_ProgressCounter) / max) * 100.0);
     QString ss = m_MessageTitle + QObject::tr(" || %1% Complete").arg(progressInt);
     notifyStatusMessage(m_MessagePrefix, m_MessageLabel, ss);
     prog += progIncrement;
