+++ Source/SIMPLib/Geometry/ImageGeom.cpp
@@ -208,7 +208,7 @@ public:
 
           // Compute the actual derivatives
           index = (z * dims[1] * dims[0]) + (y * dims[0]) + x;
-          for(int32_t i = 0; i < numComps; i++)
+          for(size_t i = 0; i < numComps; i++)
           {
             derivsPtr[index * numComps * 3 + i * 3] = xix * dValuesdXi[i] + etax * dValuesdEta[i] + zetax * dValuesdZeta[i];
 
@@ -543,9 +543,9 @@ void ImageGeom::getPlaneCoords(size_t idx, float coords[3])
 // -----------------------------------------------------------------------------
 void ImageGeom::getPlaneCoords(size_t idx[3], double coords[3])
 {
-  coords[0] = static_cast<double>(idx[0] * m_Resolution[0] + m_Origin[0]);
-  coords[1] = static_cast<double>(idx[1] * m_Resolution[1] + m_Origin[1]);
-  coords[2] = static_cast<double>(idx[2] * m_Resolution[2] + m_Origin[2]);
+  coords[0] = static_cast<double>(idx[0]) * m_Resolution[0] + m_Origin[0];
+  coords[1] = static_cast<double>(idx[1]) * m_Resolution[1] + m_Origin[1];
+  coords[2] = static_cast<double>(idx[2]) * m_Resolution[2] + m_Origin[2];
 }
 
 // -----------------------------------------------------------------------------
@@ -553,9 +553,9 @@ void ImageGeom::getPlaneCoords(size_t idx[3], double coords[3])
 // -----------------------------------------------------------------------------
 void ImageGeom::getPlaneCoords(size_t x, size_t y, size_t z, double coords[3])
 {
-  coords[0] = static_cast<double>(x * m_Resolution[0] + m_Origin[0]);
-  coords[1] = static_cast<double>(y * m_Resolution[1] + m_Origin[1]);
-  coords[2] = static_cast<double>(z * m_Resolution[2] + m_Origin[2]);
+  coords[0] = static_cast<double>(x) * m_Resolution[0] + m_Origin[0];
+  coords[1] = static_cast<double>(y) * m_Resolution[1] + m_Origin[1];
+  coords[2] = static_cast<double>(z) * m_Resolution[2] + m_Origin[2];
 }
 
 // -----------------------------------------------------------------------------
@@ -567,9 +567,9 @@ void ImageGeom::getPlaneCoords(size_t idx, double coords[3])
   size_t row = (idx / m_Dimensions[0]) % m_Dimensions[1];
   size_t plane = idx / (m_Dimensions[0] * m_Dimensions[1]);
 
-  coords[0] = static_cast<double>(column * m_Resolution[0] + m_Origin[0]);
-  coords[1] = static_cast<double>(row * m_Resolution[1] + m_Origin[1]);
-  coords[2] = static_cast<double>(plane * m_Resolution[2] + m_Origin[2]);
+  coords[0] = static_cast<double>(column) * m_Resolution[0] + m_Origin[0];
+  coords[1] = static_cast<double>(row) * m_Resolution[1] + m_Origin[1];
+  coords[2] = static_cast<double>(plane) * m_Resolution[2] + m_Origin[2];
 }
 
 // -----------------------------------------------------------------------------
@@ -611,9 +611,9 @@ void ImageGeom::getCoords(size_t idx, float coords[3])
 // -----------------------------------------------------------------------------
 void ImageGeom::getCoords(size_t idx[3], double coords[3])
 {
-  coords[0] = static_cast<double>(idx[0] * m_Resolution[0] + m_Origin[0] + (0.5f * m_Resolution[0]));
-  coords[1] = static_cast<double>(idx[1] * m_Resolution[1] + m_Origin[1] + (0.5f * m_Resolution[1]));
-  coords[2] = static_cast<double>(idx[2] * m_Resolution[2] + m_Origin[2] + (0.5f * m_Resolution[2]));
+  coords[0] = static_cast<double>(idx[0]) * m_Resolution[0] + m_Origin[0] + (0.5 * m_Resolution[0]);
+  coords[1] = static_cast<double>(idx[1]) * m_Resolution[1] + m_Origin[1] + (0.5 * m_Resolution[1]);
+  coords[2] = static_cast<double>(idx[2]) * m_Resolution[2] + m_Origin[2] + (0.5 * m_Resolution[2]);
 }
 
 // -----------------------------------------------------------------------------
@@ -621,9 +621,9 @@ void ImageGeom::getCoords(size_t idx[3], double coords[3])
 // -----------------------------------------------------------------------------
 void ImageGeom::getCoords(size_t x, size_t y, size_t z, double coords[3])
 {
-  coords[0] = static_cast<double>(x * m_Resolution[0] + m_Origin[0] + (0.5f * m_Resolution[0]));
-  coords[1] = static_cast<double>(y * m_Resolution[1] + m_Origin[1] + (0.5f * m_Resolution[1]));
-  coords[2] = static_cast<double>(z * m_Resolution[2] + m_Origin[2] + (0.5f * m_Resolution[2]));
+  coords[0] = static_cast<double>(x) * m_Resolution[0] + m_Origin[0] + (0.5 * m_Resolution[0]);
+  coords[1] = static_cast<double>(y) * m_Resolution[1] + m_Origin[1] + (0.5 * m_Resolution[1]);
+  coords[2] = static_cast<double>(z) * m_Resolution[2] + m_Origin[2] + (0.5 * m_Resolution[2]);
 }
 
 // -----------------------------------------------------------------------------
@@ -635,9 +635,9 @@ void ImageGeom::getCoords(size_t idx, double coords[3])
   size_t row = (idx / m_Dimensions[0]) % m_Dimensions[1];
   size_t plane = idx / (m_Dimensions[0] * m_Dimensions[1]);
 
-  coords[0] = static_cast<double>(column * m_Resolution[0] + m_Origin[0] + (0.5f * m_Resolution[0]));
-  coords[1] = static_cast<double>(row * m_Resolution[1] + m_Origin[1] + (0.5f * m_Resolution[1]));
-  coords[2] = static_cast<double>(plane * m_Resolution[2] + m_Origin[2] + (0.5f * m_Resolution[2]));
+  coords[0] = static_cast<double>(column) * m_Resolution[0] + m_Origin[0] + (0.5 * m_Resolution[0]);
+  coords[1] = static_cast<double>(row) * m_Resolution[1] + m_Origin[1] + (0.5 * m_Resolution[1]);
+  coords[2] = static_cast<double>(plane) * m_Resolution[2] + m_Origin[2] + (0.5 * m_Resolution[2]);
 }
 
 // -----------------------------------------------------------------------------
